feat: impact report per level
