+++ app/graphql/queries/impact.ts
@@ -1,8 +1,8 @@
 import { gql } from "@apollo/client";
 
 export const GET_IMPACT_METRICS = gql`
-  query GetImpactMetrics {
-    getImpactMetrics {
+  query GetImpactMetrics($view: VisualizationViewType!) {
+    getImpactMetrics(view: $view) {
       protected_forest
       protected_species
       removed_tons
